Use preferences in both MainActivity and TimerService
Sometimes, the MainActivity checks the preferences and then instructs
the TimerService to do something (i.e. disable WiFi and audio). Since
this is a rather roundabout way to do things, the preferences have no
gained a name and can thus be checked from both classes. In a future
commit, the TimerService can decide for itself whether to disable sound
or not.

Unfortunately, this removes the old preferences. We need either a way to
migrate the old ones, or accept that the user must set the new ones
again.