Manage session type in the TimerService
First, TimerState is split into TimerState and SessionType, as these are independent facets.

Then, the TimerService becomes more active and decides the session duration by itself based on the current SessionType. This work should rather be done here, in the class that also manages the timer, instead of the GUI, where only stuff concerning the viewing should be done. This is also important if we ever implement a widget.