Don't force app to foreground when countdown finished
On the one hand, this fixes the disappearing notification; on the other hand it improves the UX: After a session is finished, the user should be notified. The standard way for this on Android are Notiifications, so we should use those. Additionally forcing the app in the foreground is unnecessary and possibly annying.