Migrate old settings to new ones
The name of the shared preferences was changed so that both MainActivity and TimerService could load them. Now, we introduce a migration mechanism to port the old values to the new ones.

This commit also sets a settings version so that we can recognize old settings and can migrate accordingly.