Give choice of different vibration patterns
While working in a shared space, I find that the vibration pattern is
currently too long (1.5s) with impulses that are too strong (.5s each),
which may disturb coworkers.

This commit adds a new preference allowing the user to choose between
several vibration patterns. Available choices are: “strong”, the default
vibration pattern from previous commits; “soft”, a new pattern with
three short (.05s each) impulses; and “none” which disables vibration.
As this is a highly subjective matter, new patterns could be added in
the future.

The “pref_vibrate” preference is removed and replaced by a
“pref_vibration_type” preference which is mapped to the new
“VibrationType” enum’s values. In the settings screen, the switch
preference widget is replaced by a list preference for choosing among
patterns that are available.